Remove deletion until bug* is resolved
*The bug: sometimes after the PCIDevice CR is created and updated,
the PCI address field is not in the status yet. That will cause
the code to think the PCI Device is deleted, causing the CR to be
deleted. The effect can be a bunch of devices getting deleted and
recreated. Given that PCI unplugging is not a common thing, I am going
to flag this bug and prioritize other work.